+++ components/organisms/Catalog.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { cp_t } from '../../helpers/commonProps';
-import styles from './Catalog.module.css';
+import styles from '../stylesheets/Catalog.module.css';
 
 export default function Catalog(props) {
 
@@ -11,14 +11,14 @@ export default function Catalog(props) {
     },[]);
 
     async function getSheetData() {
-        let data = await fetch(`https://spreadsheets.google.com/feeds/list/1sSdxrMjaylypUCbl_vIjcMMUJkQzmd70Ks6qthrz95o/${props.t.getSheetID}/public/values?alt=json`);
+        let data = await fetch(`https://spreadsheets.google.com/feeds/list/1mkpuyaRZnHeyJtR_KuKapCImoU8cSYVJt_EUoZ5DflA/${props.t.getSheetID}/public/values?alt=json`);
         data = await data.json();
 
-        let resourceTypes = [...new Set(data.feed.entry.map(entry => entry.gsx$type.$t))];
+        let resourceTypes = [...new Set(data.feed.entry.map(entry => entry.gsx$topic.$t))];
         let sortedResources = Array.from(resourceTypes, resourceType => {
             return {
                 type: resourceType,
-                resources: data.feed.entry.filter(entry => entry.gsx$type.$t === resourceType)
+                resources: data.feed.entry.filter(entry => entry.gsx$topic.$t === resourceType)
             }
         });
         setResources(sortedResources);
@@ -28,24 +28,19 @@ export default function Catalog(props) {
 
     return (
         <div className={styles.catalog}>
-            <h2 className={styles.title}>{props.t["Open Call Catalogue"]}</h2>
+            <h2 className={styles.title}>{props.t["Parking Lot"]}</h2>
             {resources.map(resourceType =>
                 <div key={resourceType.type} className={styles.contentBox}>
                     <h3 className={styles.resourceType}>{resourceType.type}</h3>
                     <div className={styles.collection}>
                         {resourceType.resources.map((resource, index) =>
                             <div key={index} className={styles.resource}>
-                                <img src={resource.gsx$imageurl.$t} alt={resource.gsx$alttxt.$t}/>
+                                <img src={`/images/parkingLotThumbnails/${props.t["getLocale"]}/${resource.gsx$imagename.$t}`} alt={resource.gsx$alttxt.$t}/>
                                 {resource.gsx$frenchcaptions ? <figcaption>{resource.gsx$frenchcaptions.$t}</figcaption> : undefined}
-                                <h4>{resource.title.$t}</h4>
-                                <p>
-                                    <b>{props.t["Live link:"] + " "}</b>
-                                    <a href={resource.gsx$livelink.$t}>{resource.gsx$livelinktitle.$t}</a>
-                                </p>
-                                <p>
-                                    <b>{props.t["Open source code:"] + " "}</b>
-                                    <a href={resource.gsx$codelink.$t}>{resource.gsx$codelinktitle.$t}</a>
-                                </p>
+                                <a href={resource.gsx$url.$t} target="_blank">
+                                    <h4>{resource.title.$t}</h4>
+                                </a>
+                                <p>{resource.gsx$description.$t}</p>
                             </div>
                         )}
                         {3 % resourceType.resources.length != 0 ?
