+++ locales/en.js
@@ -30,20 +30,9 @@ const en = {
     "Sorting": "Sorting",
     "Most Popular": "Most Popular",
     "Recently Posted": "Recently Posted",
-    "getSheetID": "o9nsepk",
+    "getSheetID": "1",
     "getCompletedProjectsSheetID": 1,
-    "Open Call": "Open Call",
-    "Alpha": "Alpha",
-    "Learn more": "Learn more",
-    "Success stories": "Success stories",
-    "FAQs": "FAQs",
-    "Open Call Catalogue": "Open Call Catalogue",
-    "Live link:": "Live link:",
-    "Open source code:": "Open source code:",
-    "codeShareAltText": "Document with a curly bracket at the top left and a share icon at the bottom right, representing sharing.",
-    "clipboardAltText": "Clipboard featuring a list of items that are checked off.",
-    "Privacy": "Privacy",
-    "getPrivacyLink": "https://digital.canada.ca/legal/privacy/",
+    "Parking Lot": "Parking Lot",
     "browserWarningTitle": "is not compatible with this website",
     "browserWarningMessage": "To get the most out of your Busrides experience, please use one of the following supported browsers.",
     "Icons": "Icons",
