+++ locales/fr.js
@@ -30,20 +30,9 @@ const fr = {
     "Sorting": "Triage",
     "Most Popular": "Les plus populaires",
     "Recently Posted": "Publié récemment",
-    "getSheetID": "o9lw766",
+    "getSheetID": "2",
     "getCompletedProjectsSheetID": 2,
-    "Open Call": "Appel ouvert",
-    "Alpha": "Alpha",
-    "Learn more": "Plus d'information",
-    "Success stories": "Histoires de réussite",
-    "FAQs": "FAQ",
-    "Open Call Catalogue": "Le catalogue Appel ouvert",
-    "Live link:": "Lien actif :",
-    "Open source code:": "Code source libre :",
-    "codeShareAltText": "Pictogramme d’un document avec des accolades et une icône de partage, symbolisant le partage.",
-    "clipboardAltText": "Pictogramme d’un presse-papier avec une liste d’éléments cochés.",
-    "Privacy": "Confidentialité",
-    "getPrivacyLink": "https://numerique.canada.ca/transparence/confidentialite/",
+    "Parking Lot": "Parc de stationnement",
     "browserWarningTitle": "n'est pas compatible avec ce site web",
     "browserWarningMessage": "Pour tirer le meilleur parti de votre expérience Trajets en bus, veuillez utiliser l'un des navigateurs pris en charge suivants.",
     "Icons": "Icônes",
