+++ markdown/en/AboutParkingLot.md
@@ -1,0 +1,5 @@
+# Busrides Parking Lot
+
+We've got a lot of information floating around, and sometimes we just can't find a place for it all! The parking lot is a small catalogue where we can park items that need a place online.
+
+The parking lot catalogue is organized into categories so browsing is simple.
