+++ markdown/fr/AboutParkingLot.md
@@ -1,0 +1,5 @@
+# Parking trajets en bus
+
+Nous avons beaucoup d'informations qui circulent, et parfois nous n'arrivons pas à trouver une place pour tout cela ! Le parking est un petit catalogue où l'on peut garer en ligne les objets qui ont besoin d'une place.
+
+Le catalogue du parking est organisé en catégories, de sorte que la navigation est simple.
