--- openCall/components/NavButton.module.css
@@ -1,28 +1,0 @@
-.navButton {
-    background-color: #b2e3ff;
-    color: black;
-    width: auto;
-    display: inline-block;
-    padding: 10px 12px;
-}
-
-div .navButton:nth-of-type(n + 2){
-    margin-left: 10px;
-}
-
-.navButton {
-    text-decoration: none;
-    cursor: pointer;
-}
-
-.navButton p {
-    margin: 0;
-}
-
-@media only screen and (max-width: 1000px) {
-
-    .navButton {
-        margin-bottom: 15px;
-    }
-
-}
