--- openCall/markdown/en/HelpingGovernmentsRespond.md
@@ -1,5 +1,0 @@
-# Helping governments respond to COVID-19
-
-Shared problems need shared solutions. Open Call includes free digital tools to address common COVID-19 challenges and free technical support for government teams. 
-
-Browse the catalogue, or fill out our [support request form](https://forms.gle/Q7PiKNZ1tapidJuP9), and we'll help you move at the speed of need.
