--- openCall/markdown/en/HereToHelp.md
@@ -1,5 +1,0 @@
-# Pro bono tech support for your team
-
-Not a coder? No problem! Not sure what kind of resource you need, which listing is best for you, or simply need more support? Open Call can provide coaching, workshops or even hands-on development and design sprints to help you launch or improve a service. 
-
-Just fill out our [support request form](https://forms.gle/Q7PiKNZ1tapidJuP9) to get started.
