--- openCall/markdown/fr/HelpingGovernmentsRespond.md
@@ -1,5 +1,0 @@
-# Aider les administrations à lutter contre la COVID-19
-
-Les problèmes en commun demandent des solutions communes. Appel ouvert regroupe des outils numériques gratuits qui s’attaquent aux défis courants posés par la COVID-19, et offre aussi un soutien technique gratuit pour vous prêter main-forte.
-
-Parcourez le catalogue ou remplissez notre [formulaire de demande de soutien](https://forms.gle/APQHqw9cU8ZbXWa68), et nous vous aiderons à répondre aux besoins pressants des gens.
