--- openCall/markdown/fr/HereToHelp.md
@@ -1,5 +1,0 @@
-# Soutien technique gratuit pour votre équipe
-
-Vous êtes loin d’être spécialiste du code? Pas de problème! Vous ne savez pas quelle ressource s’applique à vos besoins, ou vous avez simplement besoin d’aide? Appel ouvert vous propose du coaching, des ateliers, voire des sprints pratiques de développement et de conception pour vous aider à lancer ou à améliorer un service. 
-
-Vous n’avez qu’à remplir notre [formulaire de demande de soutien](https://forms.gle/APQHqw9cU8ZbXWa68) pour commencer.
